US_62_AC1.java, US_62_AC2.java merged codes & updated
