+++ configuration.properties
@@ -1,7 +1,6 @@
 browser=chrome
-env=https://qa1.vytrack.com/user/login
-env2=https://qa2.vytrack.com/user/login
-
+#env=https://qa1.vytrack.com/user/login
+env=https://qa2.vytrack.com/user/login 
 TruckDrivers
 #KEY                  #VALUE
 usernameTruckDrivers1=User191
@@ -11,7 +10,6 @@ usernameTruckDrivers4=User194
 usernameTruckDrivers5=User195
 usernameTruckDrivers6=User196
 usernameTruckDrivers7=User197
-
 StoreManager
 #KEY                  #VALUE
 usernameStoreManager1=storemanager67
@@ -20,12 +18,10 @@ usernameStoreManager3=storemanager69
 usernameStoreManager4=storemanager70
 usernameStoreManager5=storemanager71
 usernameStoreManager6=storemanager72
-
 #KEY                  #VALUE
 SalesManager
 usernameSalesManager1=salesmanager275
 usernameSalesManager2=salesmanager276
 usernameSalesManager3=salesmanager277
 usernameSalesManager4=salesmanager278
-
 password=UserUser123
