--- src/test/java/com/vytrack/test/US_62_AC1.java
@@ -1,86 +1,0 @@
-package com.vytrack.test;
-
-import com.vytrack.utilities.BrowserUtils;
-import com.vytrack.utilities.Driver;
-import com.vytrack.utilities.VyTrack_Login;
-import org.apache.commons.io.FileUtils;
-import org.openqa.selenium.*;
-import org.testng.Assert;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Test;
-
-import java.io.File;
-import java.io.IOException;
-
-import static com.vytrack.utilities.ConfigurationReader.getProperty;
-import static com.vytrack.utilities.Driver.getDriver;
-
-public class US_62_AC1 extends TestBase {
-
-    @DataProvider(name = "provider")
-    public Object[][] credentialMethod() {
-        return new Object[][]{
-                {"usernameStoreManager1", "password"},
-                {"usernameStoreManager2", "password"},
-                {"usernameStoreManager3", "password"},
-                {"usernameStoreManager4", "password"},
-                {"usernameStoreManager5", "password"},
-                {"usernameStoreManager6", "password"},
-                {"usernameSalesManager1", "password"},
-                {"usernameSalesManager2", "password"},
-                {"usernameSalesManager3", "password"},
-                {"usernameSalesManager4", "password"},
-        };
-    }
-
-    @Test(dataProvider = "provider")
-
-    public void vyLogin(String userName, String password) {
-
-        VyTrack_Login.login(getDriver(), getProperty(userName), getProperty(password));
-
-        getDriver().findElement(By.xpath("//li[@class='dropdown dropdown-level-1'][4]")).click();
-        BrowserUtils.sleep(5);
-        getDriver().findElement(By.xpath("//span[text()='Calendar Events']")).click();
-        BrowserUtils.sleep(5);
-        getDriver().findElement(By.xpath("//a[@title='Create Calendar event']")).click();
-        BrowserUtils.sleep(5);
-        getDriver().findElement(By.xpath(" //input[@data-name='recurrence-repeat']")).click();
-        BrowserUtils.sleep(5);
-
-        WebElement repeatEveryBox = getDriver().findElement(By.xpath("//div[@class='controls recurrence-subview-control__items']//label/input[3]"));
-        String expectedRepeatNumber = "1";
-        String actualValue = repeatEveryBox.getAttribute("value");
-
-        Assert.assertEquals(actualValue, expectedRepeatNumber);
-
-    }
-
-
-
-
-        /*
-        //downcast the driver to access TakesScreenshot method
-        TakesScreenshot ts = (TakesScreenshot) driver;
-
-        //capture screenshot as output type FILE
-        File file = ts.getScreenshotAs(OutputType.FILE);
-
-        try {
-            //save the screenshot taken in destination path
-            FileUtils.copyFile(file, new File("./ScreenShot_Folder/Test1_Login.png"));
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        System.out.println("the login page screenshot is taken");
-}
-
- */
-    }
-
-
-
-
-
-
-
